+++ WowPacketParserModule.V5_4_8_18291/Parsers/QuestHandler.cs
@@ -612,7 +612,7 @@ namespace WowPacketParserModule.V5_4_8_18291.Parsers
             var guid = new byte[8];
 
             chooseReward.Item = packet.ReadUInt32("Reward Item ID");
-            chooseReward.QuestId = packet.ReadUInt32<QuestId>("Quest ID");
+            chooseReward.QuestId = (uint)packet.ReadInt32<QuestId>("Quest ID");
             guid[2] = packet.ReadBit();
             guid[6] = packet.ReadBit();
             guid[0] = packet.ReadBit();
