+++ WowPacketParserModule.V9_0_1_36216/Parsers/QuestHandler.cs
@@ -444,7 +444,7 @@ namespace WowPacketParserModule.V9_0_1_36216.Parsers
         {
             var chooseReward = packet.Holder.ClientQuestGiverChooseReward = new();
             chooseReward.QuestGiver = packet.ReadPackedGuid128("QuestGiverGUID");
-            chooseReward.QuestId = packet.ReadUInt32("QuestID");
+            chooseReward.QuestId = (uint)packet.ReadInt32("QuestID");
             chooseReward.Item = (uint)ReadRewardItem(packet, "ItemChoice").ItemID;
         }
